Newsletter + Forms: lift Gravatar into @automattic/jetpack-components
The Gravatar component (with hovercard support, SHA-256 hashing, and a
configurable defaultImage) moves into @automattic/jetpack-components
exposed at the `./gravatar` subpath. Forms switches to it; the local
copy in `projects/packages/forms/src/dashboard/components/gravatar/`
is removed and the `@gravatar-com/hovercards` and `js-sha256` deps move
with it.

Sets up Newsletter's Subscribers port to consume the same shared
component.
