+++ projects/packages/forms/changelog/use-shared-gravatar
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Forms now uses the shared `Gravatar` component from `@automattic/jetpack-components/gravatar`. No user-visible change.
