--- projects/packages/forms/src/dashboard/components/gravatar/style.scss
@@ -1,6 +1,0 @@
-.jp-forms__gravatar {
-	background-color: var(--jp-gray-5);
-	border-radius: 50%;
-	flex-shrink: 0; // Prevent shrinking when Gravatar is inside flex containers
-	overflow: hidden;
-}
